+++ pypeit/core/datacube.py
@@ -13,7 +13,7 @@ from astropy import wcs, units
 from astropy.coordinates import AltAz, SkyCoord
 from astropy.io import fits
 import scipy.optimize as opt
-from scipy.interpolate import interp1d
+from scipy.interpolate import interp1d, RegularGridInterpolator, RBFInterpolator
 import numpy as np
 
 from pypeit import msgs
@@ -23,6 +23,12 @@ from pypeit.core.flexure import calculate_image_phase
 from pypeit.core import coadd, extract, findobj_skymask, parse, skysub
 from pypeit.core.procimg import grow_mask
 from pypeit.spectrographs.util import load_spectrograph
+
+# Shapely is needed if using the resample algorithm
+try:
+    import shapely
+except ImportError:
+    shapely = None
 
 from IPython import embed
 
@@ -885,7 +891,7 @@ def generate_masterWCS(crval, cdelt, equinox=2000.0, name="Instrument Unknown"):
     # Insert the coordinate frame
     w.wcs.cname = ['RA', 'DEC', 'Wavelength']
     w.wcs.cunit = [units.degree, units.degree, units.Angstrom]
-    w.wcs.ctype = ["RA---TAN", "DEC--TAN", "AWAV"]
+    w.wcs.ctype = ["RA---TAN", "DEC--TAN", "WAVE"]
     w.wcs.crval = crval  # RA, DEC, and wavelength zeropoints
     w.wcs.crpix = [0, 0, 0]  # RA, DEC, and wavelength reference pixels
     #w.wcs.cd = np.array([[cdval[0], 0.0, 0.0], [0.0, cdval[1], 0.0], [0.0, 0.0, cdval[2]]])
@@ -989,6 +995,341 @@ def compute_weights(all_ra, all_dec, all_wave, all_sci, all_ivar, all_idx, white
     return all_wghts
 
 
+def generate_cube_resample(outfile, frame_wcs, slits, fluximg, ivarimg, raimg, decimg, waveimg, slitimg, gpm,
+                           grid_nspat=5, grid_specsep=20,
+                           overwrite=False, output_wcs=None, blaze_wave=None, blaze_spec=None, fluxcal=False,
+                           sensfunc=None, specname="PYP_SPEC", debug=False):
+    """
+    Save a datacube using the resample algorithm.
+
+    This function takes the fully calibrated input data, and resamples
+    all slices onto a regular 3D grid, while conserving flux. Note that
+    the final datacube has correlations between voxels, and this covariance
+    information is not saved.
+
+    Args:
+        outfile (`str`):
+            Filename to be used to save the datacube
+        frame_wcs (`astropy.wcs.wcs.WCS`_):
+            World coordinate system for this frame.
+        slits (:class:`pypeit.slittrace.SlitTraceSet`_)
+            Information stored about the slits
+        fluximg (`numpy.ndarray`_):
+            Surface brightness of each pixel in the frame (units = erg/s/cm^2/A/arcsec^2)
+        ivarimg (`numpy.ndarray`_):
+            Inverse variance of each pixel in the frame
+        raimg (`numpy.ndarray`_):
+            Right ascension of each pixel in the frame (units = decimal degrees)
+        decimg (`numpy.ndarray`_):
+            Declination of each pixel in the frame (units = decimal degrees)
+        waveimg (`numpy.ndarray`_):
+            Wavelength of each pixel in the frame (units = Angstroms)
+        slitimg (`numpy.ndarray`_):
+            Slit image. -1 is not on a sli, and all other
+            pixels are labelled with their spatial IDs.
+        gpm (`numpy.ndarray`_):
+            Good pixel mask (bool). True = good pixel
+        grid_nspat (int):
+            Number of grid points in the spatial direction when evaluating the
+            voxel geometry in detector coordinates. This should be an odd number
+        grid_specsep (int):
+            Number of pixels between each grid point in the spectral direction
+            when evaluating the voxel geometry in detector coordinates
+        overwrite (bool, optional):
+            If the output file exists, it will be overwritten if this parameter is True.
+        output_wcs (`astropy.wcs.wcs.WCS`_, optional):
+            World coordinate system for the output datacube. If None, frame_wcs will be used.
+        blaze_wave (`numpy.ndarray`_, optional):
+            Wavelength array of the spectral blaze function
+        blaze_spec (`numpy.ndarray`_, optional):
+            Spectral blaze function
+        fluxcal (bool, optional):
+            Are the data flux calibrated?
+        sensfunc (`numpy.ndarray`_, None, optional):
+            Sensitivity function that has been applied to the datacube
+        specname (str, optional):
+            Name of the spectrograph
+        debug (bool, optional):
+            Debug the code by writing out a residuals cube?
+    """
+    # Set the output_wcs if it's not already set
+    if output_wcs is None:
+        output_wcs = frame_wcs
+    # Check that grid_nspat is an odd number
+    if grid_nspat%2 == 0:
+        msgs.warn(f"grid_nspat must be an odd number. Using grid_nspat={grid_nspat+1} instead")
+        grid_nspat += 1
+    debug = False
+    # Get the grid spacing along the spatial direction
+    frm_cd_spat = np.sqrt(frame_wcs.wcs.cd[1, 1] ** 2 + frame_wcs.wcs.cd[0, 1] ** 2)
+    out_cd_spat = np.sqrt(output_wcs.wcs.cd[1, 1] ** 2 + output_wcs.wcs.cd[0, 1] ** 2)
+    slitlength = int(np.round(np.median(slits.get_slitlengths(initial=True, median=True))))
+    nvox_spat = int(np.ceil(slitlength*frm_cd_spat/out_cd_spat))
+    crd_vox_spat = out_cd_spat * (np.arange(nvox_spat+1) - (nvox_spat+1)// 2)  # +1 to get bin edges
+    # Get the grid spacing along the spectral direction
+    out_cr_wave = output_wcs.wcs.crval[2]
+    out_cd_wave = output_wcs.wcs.cd[2, 2]
+    nvox_wave = int(np.ceil((np.max(waveimg)-out_cr_wave)/out_cd_wave))
+    crd_vox_spec = out_cr_wave + out_cd_wave * np.arange(nvox_wave+1)  # +1 to get bin edges
+    vox_shape = (nvox_wave+1, nvox_spat+1)
+
+    # Detector spectal/spatial pixels and number of slices
+    nspec, nspat, nslice = slits.nspec, slits.nspat, slits.spat_id.size
+
+    # Generate the output datacube
+    datcube = np.zeros((nslice, nvox_spat, nvox_wave), dtype=float)
+    varcube = np.zeros((nslice, nvox_spat, nvox_wave), dtype=float)
+
+    # Transform the voxel geometry to detector pixels
+    grid_nspec = 1 + nspec // grid_specsep
+    xgrid = np.zeros((grid_nspec, grid_nspat), dtype=int)
+    ygridt = np.zeros(grid_nspec, dtype=int)
+    ygridt[-1] = nspec - 1
+    ygridt[1:-1] = (nspec % grid_specsep + 2 * grid_specsep) // 2 + np.arange(grid_nspec - 2) * grid_specsep
+    ygrid = ygridt[:, np.newaxis].repeat(grid_nspat, axis=1)
+    ra0, dec0 = np.zeros(nslice), np.zeros(nslice)
+    offsimg = np.zeros_like(waveimg)
+    varimgsq = utils.inverse(ivarimg ** 2)
+    import time
+    atime = time.time()
+    for sl, spat_id in enumerate(slits.spat_id):
+        msgs.info(f"Calculating voxel geometry for slit {spat_id} -- {(time.time()-atime)/60}")
+        # Calculate RA and Dec of central traces
+        wsl = np.where(slitimg == spat_id)
+        this_ra, this_dec, this_wave = raimg[wsl], decimg[wsl], waveimg[wsl]
+        _, spat_posn, _ = frame_wcs.wcs_world2pix(this_ra, this_dec, this_wave*1.0E-10, 0)
+        asrt = np.argsort(spat_posn)
+        ra0[sl] = np.interp(0.0, spat_posn[asrt], this_ra[asrt])
+        dec0[sl] = np.interp(0.0, spat_posn[asrt], this_dec[asrt])
+        # Generate the offsets
+        cosdec = np.cos(dec0[sl] * np.pi / 180.0)
+        diff_ra, diff_dec = (this_ra - ra0[sl]) * cosdec, this_dec - dec0[sl]
+        msgs.bug("There is sometimes a sign error that needs to be resolved here...")
+        msgs.error("Use another algorithm for the time being...")
+        if np.max(diff_ra)-np.min(diff_ra) > np.max(diff_dec)-np.min(diff_dec):
+            sgn = np.sign(diff_ra)
+        else:
+            sgn = np.sign(diff_dec)
+        offsimg[wsl] = -sgn * np.sqrt(diff_ra**2 + diff_dec**2)
+        # Update the xgrid values for this slice
+        for yy in range(grid_nspec):
+            wsl = np.where(slitimg == spat_id)
+            allind = wsl[1][np.where(wsl[0] == ygridt[yy])]
+            xgrid[yy, 0] = np.min(allind)
+            xgrid[yy, -1] = np.max(allind)
+            numpix = xgrid[yy, -1] - xgrid[yy, 0]
+            sep = numpix // (grid_nspat - 1)
+            xgrid[yy, 1:-1] = xgrid[yy, 0] + (numpix % sep + 2 * sep) // 2 + np.arange(grid_nspat - 2) * sep
+        # Extract offset + wavelength information and estimate transform
+        grid_coord = (ygrid.flatten(), xgrid.flatten())
+        grid_offs = offsimg[grid_coord]
+        grid_wave = waveimg[grid_coord]
+        src = np.column_stack((grid_wave, grid_offs))
+        dst = np.column_stack(grid_coord).astype(float)
+        # Transform the voxel coordinates to detector coordinates
+        evalpos = np.column_stack((crd_vox_spec[:,np.newaxis].repeat(crd_vox_spat.size, axis=1).flatten(),
+                                   crd_vox_spat[np.newaxis,:].repeat(crd_vox_spec.size, axis=0).flatten()))
+        # tform = LinearNDInterpolator(src, dst, rescale=True)
+        # crd_det_tmp = tform(evalpos)
+
+        src_off = np.min(src, axis=0)
+        src_scl = np.max(src-src_off, axis=0)
+        dst_off = np.min(dst, axis=0)
+        dst_scl = np.max(dst-dst_off, axis=0)
+        tform = RBFInterpolator((src-src_off)/src_scl, (dst-dst_off)/dst_scl, smoothing=0.01)
+        crd_det = dst_off + dst_scl * tform((evalpos-src_off)/src_scl)
+        if debug:
+            plt.plot(crd_det[:, 0], crd_det[:, 1], 'rx')
+            #plt.plot(crd_det_tmp[:, 0], crd_det_tmp[:, 1], 'bx')
+            plt.plot(np.arange(slits.left_init.shape[0]), slits.left_init[:, 0], 'k-')
+            plt.plot(np.arange(slits.right_init.shape[0]), slits.right_init[:, 0], 'k-')
+            plt.show()
+
+    # Calculate an "offsets" image, which indicates the offset in arcsec from (RA_0, DEC_0)
+    # Create two splines of the offsets image: (1) offset predicts RA; (2) offset predicts Dec.
+    # Use these splines to calculate the RA and DEC of the voxels, combine this with the output wavelength grid.
+    # Generate all RA, DEC, WAVELENGTH triples (i.e. find the RA,DEC pairs along constant wavelength, for all wavelengths)
+    # Use the WCS (which contains the astrometric transform) to go from world to pix
+    #    i.e. need to invert this:
+    #    world_ra, world_dec, _ = wcs.wcs_pix2world(slitID, evalpos, tilts[onslit_init]*(nspec-1), 0)
+    # This gives us the x,y detector positions of the voxel geometry
+        from shapely.geometry import Polygon, box as shapelyBox
+        from shapely.strtree import STRtree
+
+        crd_det_spec, crd_det_spat = crd_det[:, 0].reshape(vox_shape), crd_det[:, 1].reshape(vox_shape)
+        # Generate a list of all detector pixels in this slice
+        detpix_polys = []
+        pix_spec, pix_spat = np.where(slitimg == spat_id)
+        for ss in range(pix_spat.size):
+            detpix_polys.append(shapely.geometry.box(pix_spat[ss], pix_spec[ss], pix_spat[ss]+1, pix_spec[ss]+1))
+        # Create a Sort-Tile-Recursive tree of the detector pixels to quickly query overlapping voxels
+        detgeom = shapely.strtree.STRtree(detpix_polys)
+        # Loop through all voxels for this slice and calculate the overlapping area
+        #all_area = np.zeros_like(fluximg)
+        #atime=time.time()
+        for wv in range(nvox_wave):
+            for sp in range(nvox_spat):
+                # Generate the voxel coordinates in detector pixel space (points must be counter-clockwise)
+                voxel_geom = shapely.geometry.Polygon([(crd_det_spat[wv, sp],   crd_det_spec[wv,   sp]),
+                                                       (crd_det_spat[wv, sp+1], crd_det_spec[wv,   sp]),
+                                                       (crd_det_spat[wv, sp+1], crd_det_spec[wv+1, sp]),
+                                                       (crd_det_spat[wv, sp],   crd_det_spec[wv+1, sp]),
+                                                       (crd_det_spat[wv, sp],   crd_det_spec[wv,   sp])])
+                # Find overlapping detector pixels
+                result = detgeom.query(voxel_geom)
+                # Sum all overlapping flux-weighted areas
+                this_flx = 0
+                this_var = 0
+                this_area = 0
+                for pp in range(len(result)):
+                    area = voxel_geom.intersection(result[pp]).area
+                    pix_spat = int(min(result[pp].exterior.coords[0][0], result[pp].exterior.coords[2][0]))
+                    pix_spec = int(min(result[pp].exterior.coords[0][1], result[pp].exterior.coords[2][1]))
+                    if ivarimg[pix_spec, pix_spat] != 0.0:
+                        this_flx += area * fluximg[pix_spec, pix_spat]
+                        this_var += area**2 * varimgsq[pix_spec, pix_spat]
+                        this_area += area
+                    #all_area[pix_spec, pix_spat] += area
+                # Fill in the datacube
+                this_area = 1 if this_area == 0 else this_area
+                datcube[sl, sp, wv] = this_flx / this_area
+                varcube[sl, sp, wv] = this_var / this_area**2
+
+    # Generate a header
+    hdr = output_wcs.to_header()
+
+    # Add the unit of flux to the header
+    if fluxcal:
+        hdr['FLUXUNIT'] = (PYPEIT_FLUX_SCALE, "Flux units -- erg/s/cm^2/Angstrom/arcsec^2")
+    else:
+        hdr['FLUXUNIT'] = (1, "Flux units -- counts/s/Angstrom/arcsec^2")
+
+    # Save the final datacube
+    msgs.info("Saving datacube as: {0:s}".format(outfile))
+    final_cube = DataCube(datcube.T, varcube.T, specname, blaze_wave, blaze_spec, sensfunc=sensfunc, fluxed=fluxcal)
+    final_cube.to_file(outfile, hdr=hdr, overwrite=overwrite)
+
+def generate_cube_subsample(outfile, output_wcs, all_sci, all_ivar, all_wghts, all_wave, tilts, slits, slitid_img_gpm,
+                            astrom_trans, bins, subsample=10, overwrite=False, blaze_wave=None, blaze_spec=None,
+                            fluxcal=False, sensfunc=None, specname="PYP_SPEC", debug=False):
+    """
+    Save a datacube using the subsample algorithm. This algorithm splits
+    each detector pixel into multiple subpixels, and then assigns each
+    subpixel to a voxel. For example, if subsample=10, then each detector
+    pixel is subsampled by 10^2=100 subpixels. When subsample=1, this
+    corresponds to the nearest grid point (NGP) algorithm.
+
+    Args:
+        outfile (`str`):
+            Filename to be used to save the datacube
+        output_wcs (`astropy.wcs.wcs.WCS`_):
+            Output world coordinate system.
+        all_sci (`numpy.ndarray`_):
+            1D flattened array containing the counts of each pixel from all spec2d files
+        all_ivar (`numpy.ndarray`_):
+            1D flattened array containing the inverse variance of each pixel from all spec2d files
+        all_wghts (`numpy.ndarray`_):
+            1D flattened array containing the weights of each pixel to be used in the combination
+        all_wave (`numpy.ndarray`_)
+            1D flattened array containing the wavelength of each pixel (units = Angstroms)
+        tilts (`numpy.ndarray`_)
+            2D wavelength tilts frame
+        slits (:class:`pypeit.slittrace.SlitTraceSet`_)
+            Information stored about the slits
+        slitid_img_gpm (`numpy.ndarray`_)
+            An image indicating which pixels belong to a slit (0 = not on a slit or a masked pixel).
+            Any positive value indicates the spatial ID of the pixel.
+        astrom_trans (:class:`pypeit.alignframe.AlignmentSplines`_):
+            A Class containing the transformation between detector pixel coordinates and WCS pixel coordinates
+        bins (tuple):
+            A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates
+        subsample (`int`, optional):
+            What is the subsampling factor. Higher values give more reliable results, but note
+            that the time required goes as N^2. The default value is 10, which subsamples each detector pixel into
+            100 subpixels (i.e. 10^2).
+        overwrite (`bool`, optional):
+            If True, the output cube will be overwritten.
+        blaze_wave (`numpy.ndarray`_, optional):
+            Wavelength array of the spectral blaze function
+        blaze_spec (`numpy.ndarray`_, optional):
+            Spectral blaze function
+        fluxcal (bool, optional):
+            Are the data flux calibrated?
+        sensfunc (`numpy.ndarray`_, None, optional):
+            Sensitivity function that has been applied to the datacube
+        specname (str, optional):
+            Name of the spectrograph
+        debug (bool):
+            If True, a residuals cube will be output. If the datacube generation is correct, the
+            distribution of pixels in the residual cube with no flux should have mean=0 and std=1.
+    """
+    # Prepare the output arrays
+    outshape = (bins[0].size-1, bins[1].size-1, bins[2].size-1)
+    datacube, varcube, normcube = np.zeros(outshape), np.zeros(outshape), np.zeros(outshape)
+    if debug: residcube = np.zeros(outshape)
+    # Subsample each pixel
+    ssamp_offs = np.arange(0.5/subsample, 1, 1/subsample) - 0.5  # -0.5 is to offset from the centre of each pixel.
+    area = 1/subsample**2
+    all_wght_subsmp = all_wghts * area
+    # Loop through all slits
+    all_sltid = slitid_img_gpm[(slitid_img_gpm > 0)]
+    all_var = utils.inverse(all_ivar)
+    wave0, wave_delta = output_wcs.wcs.crval[2], output_wcs.wcs.cd[2, 2]
+    for sl, spatid in enumerate(slits.spat_id):
+        msgs.info(f"Resampling slit {sl+1}/{slits.nslits} into the datacube")
+        this_sl = np.where(all_sltid==spatid)
+        wpix = np.where(slitid_img_gpm==spatid)
+        slitID = np.ones(wpix[0].size) * sl - output_wcs.wcs.crpix[0]
+        # Generate a spline between spectral pixel position and wavelength
+        yspl = tilts[wpix]*(slits.nspec - 1)
+        wspl = all_wave[this_sl]
+        asrt = np.argsort(yspl)
+        wave_spl = interp1d(yspl[asrt], wspl[asrt], kind='linear', bounds_error=False, fill_value='extrapolate')
+        for xx in range(subsample):
+            for yy in range(subsample):
+                # Calculate the tranformation from detector pixels to voxels
+                spatpos = astrom_trans.transform(sl, wpix[1] + ssamp_offs[xx], wpix[0] + ssamp_offs[yy])
+                # TODO :: The tilts in the following line is evaluated at the pixel location, not the subsampled pixel location
+                # A simple fix is implemented for the spectral direction, but this is not so straightforward for the spatial direction
+                # Probably, the correction in the spatial direction is so tiny, that this doesn't matter...
+                specpos = (wave_spl(tilts[wpix]*(slits.nspec - 1) + ssamp_offs[yy]) - wave0) / wave_delta
+                # Now assemble this position of the datacube
+                pix_coord = np.column_stack((slitID, spatpos, specpos))
+                tmp_dc, _ = np.histogramdd(pix_coord, bins=bins, weights=all_sci[this_sl] * all_wght_subsmp[this_sl])
+                tmp_vr, _ = np.histogramdd(pix_coord, bins=bins, weights=all_var[this_sl] * all_wght_subsmp[this_sl]**2)
+                tmp_nm, _ = np.histogramdd(pix_coord, bins=bins, weights=all_wght_subsmp[this_sl])
+                if debug:
+                    tmp_rsd, _ = np.histogramdd(pix_coord, bins=bins, weights=all_sci[this_sl] * np.sqrt(all_ivar[this_sl]))
+                    residcube += tmp_rsd
+                datacube += tmp_dc
+                varcube += tmp_vr
+                normcube += tmp_nm
+    # Normalise the datacube and variance cube
+    nc_inverse = utils.inverse(normcube)
+    datacube *= nc_inverse
+    varcube *= nc_inverse**2
+    if debug: residcube *= nc_inverse
+
+    # Prepare the header, and add the unit of flux to the header
+    hdr = output_wcs.to_header()
+    if fluxcal:
+        hdr['FLUXUNIT'] = (PYPEIT_FLUX_SCALE, "Flux units -- erg/s/cm^2/Angstrom/arcsec^2")
+    else:
+        hdr['FLUXUNIT'] = (1, "Flux units -- counts/s/Angstrom/arcsec^2")
+
+    # Write out the datacube
+    msgs.info("Saving datacube as: {0:s}".format(outfile))
+    final_cube = DataCube(datacube.T, varcube.T, specname, blaze_wave, blaze_spec, sensfunc=sensfunc, fluxed=fluxcal)
+    final_cube.to_file(outfile, hdr=hdr, overwrite=overwrite)
+
+    # Save a residuals cube, if requested
+    if debug:
+        outfile_resid = outfile.replace(".fits", "_resid.fits")
+        msgs.info("Saving residuals datacube as: {0:s}".format(outfile_resid))
+        hdu = fits.PrimaryHDU(residcube.T, header=hdr)
+        hdu.writeto(outfile_resid, overwrite=overwrite)
+
+
 def generate_cube_ngp(outfile, hdr, all_sci, all_ivar, all_wghts, pix_coord, bins,
                       overwrite=False, blaze_wave=None, blaze_spec=None, fluxcal=False,
                       sensfunc=None, specname="PYP_SPEC", debug=False):
@@ -1011,15 +1352,17 @@ def generate_cube_ngp(outfile, hdr, all_sci, all_ivar, all_wghts, pix_coord, bin
             pixel in the processed spec2d frames. After setting up an astropy WCS, pix_coord is
             returned by the function: `astropy.wcs.WCS.wcs_world2pix_`
         bins (tuple):
-            A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates    :param overwrite:
+            A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates
+        overwrite (`bool`):
+            If True, the output cube will be overwritten.
         blaze_wave (`numpy.ndarray`_):
             Wavelength array of the spectral blaze function
         blaze_spec (`numpy.ndarray`_):
             Spectral blaze function
-        sensfunc (`numpy.ndarray`_, None):
-            Sensitivity function that has been applied to the datacube
         fluxcal (bool):
             Are the data flux calibrated?
+        sensfunc (`numpy.ndarray`_, None):
+            Sensitivity function that has been applied to the datacube
         specname (str):
             Name of the spectrograph
         debug (bool):
@@ -1128,13 +1471,30 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
     flatpar = parset['calibrations']['flatfield']
     senspar = parset['sensfunc']
 
-    # Get the detector number and string representation
-    det = 1 if parset['rdx']['detnum'] is None else parset['rdx']['detnum']
-    detname = spec.get_det_name(det)
-
     # prep
     numfiles = len(files)
     combine = cubepar['combine']
+    method = cubepar['method'].lower()
+
+    # Determine what method is requested
+    if method == "resample":
+        if shapely is None:
+            msgs.error("To use the 'resample' algorithm to make your datacube, you need to install shapely.")
+        else:
+            msgs.info("Adopting the 'resample' algorithm to generate the datacube.")
+            if combine:
+                msgs.warn("Cannot combine cubes with the 'resample' algorithm - generating individual datacubes.")
+                combine = False
+    elif method == "subsample":
+        msgs.info("Adopting the subsample algorithm to generate the datacube.")
+    elif method == "ngp":
+        msgs.info("Adopting the nearest grid point (NGP) algorithm to generate the datacube.")
+    else:
+        msgs.error(f"The following datacube method is not allowed: {method}")
+
+    # Get the detector number and string representation
+    det = 1 if parset['rdx']['detnum'] is None else parset['rdx']['detnum']
+    detname = spec.get_det_name(det)
 
     # Check if the output file exists
     if combine:
@@ -1338,6 +1698,20 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
         # Grab the slit edges
         slits = spec2DObj.slits
 
+        # debug = False
+        # if debug:
+        #     embed()
+        #     assert(False)
+        #     from matplotlib import pyplot as plt
+        #     colors = plt.cm.jet(np.linspace(0, 1, slits.nslits))
+        #     left, right, _ = slits.select_edges(initial=True, flexure=None)
+        #     mid = np.round(0.5*(left+right)).astype(np.int)
+        #     for ss in range(slits.nslits):
+        #         wv = waveimg[(np.arange(waveimg.shape[0]),mid[:,ss],)]
+        #         fx = sciimg[(np.arange(waveimg.shape[0]),mid[:,ss],)]
+        #         plt.plot(wv,fx, color=colors[ss])
+        #     plt.show()
+
         wave0 = waveimg[waveimg != 0.0].min()
         # Calculate the delta wave in every pixel on the slit
         waveimp = np.roll(waveimg, 1, axis=0)
@@ -1393,9 +1767,8 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
             msgs.warn("Spatial scale requested ({0:f} arcsec) is less than the slicer scale ({1:f} arcsec)".format(3600.0*dspat, 3600.0*slscl))
 
         # Loading the alignments frame for these data
-        astrometric = cubepar['astrometric']
         alignments = None
-        if astrometric:
+        if cubepar['astrometric']:
             alignfile = masterframe.construct_file_name(alignframe.Alignments, hdr['TRACMKEY'],
                                                         master_dir=hdr['PYPMFDIR'])
             if os.path.exists(alignfile) and cubepar['astrometric']:
@@ -1404,14 +1777,19 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
             else:
                 msgs.warn("Could not find Master Alignment frame:"+msgs.newline()+alignfile)
                 msgs.warn("Astrometric correction will not be performed")
-                astrometric = False
         else:
             msgs.info("Astrometric correction will not be performed")
-
+        # If nothing better was provided, use the slit edges
+        if alignments is None:
+            left, right, _ = slits.select_edges(initial=True, flexure=flexure)
+            locations = [0.0, 1.0]
+            traces = np.append(left[:,None,:], right[:,None,:], axis=1)
+        else:
+            traces = alignments.traces
         # Generate an RA/DEC image
         msgs.info("Generating RA/DEC image")
-        raimg, decimg, minmax = slits.get_radec_image(frame_wcs, alignments, spec2DObj.tilts, locations,
-                                                      astrometric=astrometric, initial=True, flexure=flexure)
+        raimg, decimg, minmax, ast_trans = slits.get_radec_image(frame_wcs, traces, spec2DObj.tilts, locations,
+                                                                 initial=True, flexure=flexure)
 
         # Perform the DAR correction
         if wave_ref is None:
@@ -1524,22 +1902,46 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
                 outfile = get_output_filename(fil, cubepar['output_filename'], combine, ff+1)
             # Generate individual whitelight images of each spec2d file
             if cubepar['save_whitelight']:
+                if method == 'resample':
+                    # TODO :: Implement this feature... May be better to generate this after making the cube?
+                    msgs.warn("Whitelight images are not implemented with the 'resample' algorithm.")
+                    msgs.info("Generating a whitelight image with the NGP algorithm.")
                 out_whitelight = os.path.splitext(outfile)[0] + "_whitelight.fits"
                 whitelight_img, _, wlwcs = make_whitelight_frompixels(raimg[onslit_gpm], decimg[onslit_gpm], wave_ext,
                                                                       flux_sav[resrt], np.ones(numpix), np.zeros(numpix), dspat)
                 msgs.info("Saving white light image as: {0:s}".format(out_whitelight))
                 img_hdu = fits.PrimaryHDU(whitelight_img.T, header=wlwcs.to_header())
                 img_hdu.writeto(out_whitelight, overwrite=overwrite)
-            # Make the datacube
+            # Get the coordinate bounds
             slitlength = int(np.round(np.median(slits.get_slitlengths(initial=True, median=True))))
             numwav = int((np.max(waveimg) - wave0) / dwv)
             bins = spec.get_datacube_bins(slitlength, minmax, numwav)
-            msgs.info("Generating pixel coordinates")
-            pix_coord = frame_wcs.wcs_world2pix(np.vstack((raimg[onslit_gpm], decimg[onslit_gpm], wave_ext * 1.0E-10)).T, 0)
-            hdr = frame_wcs.to_header()
-            generate_cube_ngp(outfile, hdr, flux_sav[resrt], ivar_sav[resrt], np.ones(numpix), pix_coord, bins,
-                              overwrite=overwrite, blaze_wave=blaze_wave, blaze_spec=blaze_spec,
-                              fluxcal=fluxcal, specname=specname)
+            # Generate the output WCS for the datacube
+            crval_wv = cubepar['wave_min'] if cubepar['wave_min'] is not None else 1.0E10 * frame_wcs.wcs.crval[2]
+            cd_wv = cubepar['wave_delta'] if cubepar['wave_delta'] is not None else 1.0E10 * frame_wcs.wcs.cd[2, 2]
+            cd_spat = cubepar['spatial_delta'] if cubepar['spatial_delta'] is not None else px_deg * 3600.0
+            output_wcs = spec.get_wcs(spec2DObj.head0, slits, detector.platescale, crval_wv, cd_wv, spatial_scale=cd_spat)
+            # Make the datacube
+            if method in ['subsample', 'ngp']:
+                if method == 'ngp': subsample = 1
+                else: subsample = cubepar['subsample']
+                # Get the slit image and then unset pixels in the slit image that are bad
+                slitid_img_gpm = slitid_img_init.copy()
+                slitid_img_gpm[(bpmmask != 0) | (~sky_is_good)] = 0
+                generate_cube_subsample(outfile, output_wcs, flux_sav[resrt], ivar_sav[resrt], np.ones(numpix),
+                                        wave_ext, spec2DObj.tilts, slits, slitid_img_gpm, ast_trans, bins,
+                                        overwrite=overwrite, blaze_wave=blaze_wave, blaze_spec=blaze_spec,
+                                        fluxcal=fluxcal, specname=specname, subsample=subsample)
+            elif method == 'resample':
+                fluximg, ivarimg = np.zeros_like(raimg), np.zeros_like(raimg)
+                fluximg[onslit_gpm] = flux_sav[resrt]
+                ivarimg[onslit_gpm] = ivar_sav[resrt]
+                # Now generate the cube
+                generate_cube_resample(outfile, frame_wcs, slits, fluximg, ivarimg, raimg, decimg, waveimg, slitid_img_init, onslit_gpm,
+                                       overwrite=overwrite, output_wcs=output_wcs, blaze_wave=blaze_wave, blaze_spec=blaze_spec,
+                                       fluxcal=fluxcal, specname=specname)
+            else:
+                msgs.error(f"The following method is not yet implemented: {method}")
             continue
 
         # Store the information
@@ -1649,12 +2051,12 @@ def coadd_cube(files, opts, spectrograph=None, parset=None, overwrite=False):
     coord_min = [ra_min, dec_min, wav_min]
     coord_dlt = [dspat, dspat, dwv]
     masterwcs = generate_masterWCS(coord_min, coord_dlt, name=specname)
-    msgs.info(msgs.newline()+"-"*40 +
+    msgs.info(msgs.newline() + "-" * 40 +
               msgs.newline() + "Parameters of the WCS:" +
               msgs.newline() + "RA   min, max = {0:f}, {1:f}".format(ra_min, ra_max) +
               msgs.newline() + "DEC  min, max = {0:f}, {1:f}".format(dec_min, dec_max) +
               msgs.newline() + "WAVE min, max = {0:f}, {1:f}".format(wav_min, wav_max) +
-              msgs.newline() + "Spaxel size = {0:f}''".format(3600.0*dspat) +
+              msgs.newline() + "Spaxel size = {0:f} arcsec".format(3600.0*dspat) +
               msgs.newline() + "Wavelength step = {0:f} A".format(dwv) +
               msgs.newline() + "-" * 40)